+++ benchmark/data/cpp_code_test/sign_12_3_128.cpp
@@ -9,9 +9,9 @@ float signf(float input) {
   }
 }
 extern "C" void sign_kernel(float *input, float *output) {
-  for (size_t i = 0; i < 12; i++) {
-    for (size_t j = 0; j < 3; j++) {
-      for (size_t k = 0; k < 128; k++) {
+  for (int i = 0; i < 12; i++) {
+    for (int j = 0; j < 3; j++) {
+      for (int k = 0; k < 128; k++) {
         output[i * 3 * 128 + j * 128 + k] =
             signf(input[i * 3 * 128 + j * 128 + k]);
       }
