+++ benchmark/data/cpp_code_test/sign_45_25.cpp
@@ -9,8 +9,8 @@ float signf(float input) {
   }
 }
 extern "C" void sign_kernel(float *input, float *output) {
-  for (size_t i = 0; i < 45; i++) {
-    for (size_t j = 0; j < 25; j++) {
+  for (int i = 0; i < 45; i++) {
+    for (int j = 0; j < 25; j++) {
       output[i * 25 + j] = signf(input[i * 25 + j]);
     }
   }
