+++ benchmark/data/cpp_code_test/sign_5_128.cpp
@@ -9,8 +9,8 @@ float signf(float input) {
   }
 }
 extern "C" void sign_kernel(float *input, float *output) {
-  for (size_t i = 0; i < 5; i++) {
-    for (size_t j = 0; j < 128; j++) {
+  for (int i = 0; i < 5; i++) {
+    for (int j = 0; j < 128; j++) {
       output[i * 128 + j] = signf(input[i * 128 + j]);
     }
   }
