+++ benchmark/data/cpp_code_test/sign_5_12_23_128.cpp
@@ -9,10 +9,10 @@ float signf(float input) {
   }
 }
 extern "C" void sign_kernel(float *input, float *output) {
-  for (size_t i = 0; i < 5; i++) {
-    for (size_t j = 0; j < 12; j++) {
-      for (size_t k = 0; k < 23; k++) {
-        for (size_t l = 0; l < 128; l++) {
+  for (int i = 0; i < 5; i++) {
+    for (int j = 0; j < 12; j++) {
+      for (int k = 0; k < 23; k++) {
+        for (int l = 0; l < 128; l++) {
           output[i * 12 * 23 * 128 + j * 23 * 128 + k * 128 + l] =
               signf(input[i * 12 * 23 * 128 + j * 23 * 128 + k * 128 + l]);
         }
