+++ benchmark/data/cpp_code_test/sign_7_1_6_7.cpp
@@ -9,10 +9,10 @@ float signf(float input) {
   }
 }
 extern "C" void sign_kernel(float *input, float *output) {
-  for (size_t i = 0; i < 7; i++) {
-    for (size_t j = 0; j < 1; j++) {
-      for (size_t k = 0; k < 6; k++) {
-        for (size_t l = 0; l < 7; l++) {
+  for (int i = 0; i < 7; i++) {
+    for (int j = 0; j < 1; j++) {
+      for (int k = 0; k < 6; k++) {
+        for (int l = 0; l < 7; l++) {
           output[i * 1 * 6 * 7 + j * 6 * 7 + k * 7 + l] =
               signf(input[i * 1 * 6 * 7 + j * 6 * 7 + k * 7 + l]);
         }
