+++ benchmark/data/dlboost_code_test/relu_3_4_5.cpp
@@ -1,9 +1,9 @@
 
 float reluf(float input) { return input > 0 ? input : 0; }
 extern "C" void relu_kernel(float *input, float *output) {
-  for (size_t i = 0; i < 3; i++) {
-    for (size_t j = 0; j < 4; j++) {
-      for (size_t k = 0; k < 5; k++) {
+  for (int i = 0; i < 3; i++) {
+    for (int j = 0; j < 4; j++) {
+      for (int k = 0; k < 5; k++) {
         output[i * 4 * 5 + j * 5 + k] = reluf(input[i * 4 * 5 + j * 5 + k]);
       }
     }
