+++ benchmark/data/dlboost_code_test/relu_45_25.cpp
@@ -1,8 +1,8 @@
 
 float reluf(float input) { return input > 0 ? input : 0; }
 extern "C" void relu_kernel(float *input, float *output) {
-  for (size_t i = 0; i < 45; i++) {
-    for (size_t j = 0; j < 25; j++) {
+  for (int i = 0; i < 45; i++) {
+    for (int j = 0; j < 25; j++) {
       output[i * 25 + j] = reluf(input[i * 25 + j]);
     }
   }
