+++ benchmark/data/dlboost_code_test/relu_5_128.cpp
@@ -1,8 +1,8 @@
 
 float reluf(float input) { return input > 0 ? input : 0; }
 extern "C" void relu_kernel(float *input, float *output) {
-  for (size_t i = 0; i < 5; i++) {
-    for (size_t j = 0; j < 128; j++) {
+  for (int i = 0; i < 5; i++) {
+    for (int j = 0; j < 128; j++) {
       output[i * 128 + j] = reluf(input[i * 128 + j]);
     }
   }
