+++ benchmark/data/dlboost_code_test/sign_8_10_64.cpp
@@ -9,9 +9,9 @@ float signf(float input) {
   }
 }
 extern "C" void sign_kernel(float *input, float *output) {
-  for (size_t i = 0; i < 8; i++) {
-    for (size_t j = 0; j < 10; j++) {
-      for (size_t k = 0; k < 64; k++) {
+  for (int i = 0; i < 8; i++) {
+    for (int j = 0; j < 10; j++) {
+      for (int k = 0; k < 64; k++) {
         output[i * 10 * 64 + j * 64 + k] =
             signf(input[i * 10 * 64 + j * 64 + k]);
       }
